Renamed the MainPageCard component to PaintingCard. Now the Picture component is used inside the PaintingCard for adaptive image output
