--- src/components/MainPageCard/MainPageCard.props.ts
@@ -1,6 +1,0 @@
-export interface MainPageCardProps {
-  imgSrc: string;
-  name: string;
-  years: string;
-  isDarkTheme?: boolean;
-}
