--- src/components/MainPageCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MainPageCard';
