+++ src/components/PaintingCard/PaintingCard.stories.tsx
@@ -1,11 +1,11 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import MainPageCard from './index';
+import PaintingCard from './index';
 
-const meta: Meta<typeof MainPageCard> = {
-  component: MainPageCard,
+const meta: Meta<typeof PaintingCard> = {
+  component: PaintingCard,
 };
 
-type Story = StoryObj<typeof MainPageCard>;
+type Story = StoryObj<typeof PaintingCard>;
 
 const Default: Story = {
   args: {
